# Speedup RGB565 rendering