--- a/internal/backends/winit/renderer/sw.rs
+++ b/internal/backends/winit/renderer/sw.rs
@@ -50,6 +50,8 @@ impl From<PremultipliedRgbaColor> for SoftBufferPixel {
 }
 
 impl TargetPixel for SoftBufferPixel {
+    type BlendableColor = PremultipliedRgbaColor;
+
     fn blend(&mut self, color: PremultipliedRgbaColor) {
         let mut x = PremultipliedRgbaColor::from(*self);
         x.blend(color);