By not going through a RGB888 premultiplied color, but using a special encoded format for premultiplied color so we can operate on all color at once in a u32


Unfortunately, this is a breaking change as it changes the TargetPixel trait.